refactor: merge recurring classes validation tests into single file
- Consolidate validation tests from recurringClasses.test.ts into recurringClass.test.ts
- Remove duplicate test file to maintain single source of truth
- Add validation tests for all endpoints:
  - GET /recurring-classes (query validation)
  - GET /recurring-classes/:id (param validation)
  - GET /recurring-classes/by-instructorId (query validation)
  - POST /recurring-classes (body validation)
  - PUT /recurring-classes/:id (param + body validation)
- Keep existing business logic tests in same file
- All 147 tests passing
